refactor: Split pipeline into separate stages
