+++ server/engine/src/main/java/org/eclipse/lsp/cobol/core/engine/CobolLanguageEngine.java
@@ -14,20 +14,13 @@
  */
 package org.eclipse.lsp.cobol.core.engine;
 
+import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
-import org.antlr.v4.runtime.CharStreams;
-import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.tree.ParseTreeListener;
-import org.eclipse.lsp.cobol.common.AnalysisConfig;
-import org.eclipse.lsp.cobol.common.AnalysisResult;
-import org.eclipse.lsp.cobol.common.ResultWithErrors;
-import org.eclipse.lsp.cobol.common.SubroutineService;
-import org.eclipse.lsp.cobol.common.copybook.CopybookConfig;
-import org.eclipse.lsp.cobol.common.dialects.DialectOutcome;
-import org.eclipse.lsp.cobol.common.dialects.DialectProcessingContext;
+import org.eclipse.lsp.cobol.common.*;
 import org.eclipse.lsp.cobol.common.error.ErrorSeverity;
 import org.eclipse.lsp.cobol.common.error.SyntaxError;
 import org.eclipse.lsp.cobol.common.error.ErrorCodes;
@@ -35,35 +28,18 @@ import org.eclipse.lsp.cobol.common.mapping.ExtendedSource;
 import org.eclipse.lsp.cobol.common.mapping.OriginalLocation;
 import org.eclipse.lsp.cobol.common.mapping.TextTransformations;
 import org.eclipse.lsp.cobol.common.message.MessageService;
-import org.eclipse.lsp.cobol.common.model.tree.*;
-import org.eclipse.lsp.cobol.common.model.tree.variable.*;
-import org.eclipse.lsp.cobol.common.processor.ProcessingContext;
-import org.eclipse.lsp.cobol.common.processor.ProcessingPhase;
-import org.eclipse.lsp.cobol.common.processor.ProcessorDescription;
-import org.eclipse.lsp.cobol.common.utils.RangeUtils;
 import org.eclipse.lsp.cobol.common.utils.ThreadInterruptionUtil;
-import org.eclipse.lsp.cobol.core.CobolLexer;
-import org.eclipse.lsp.cobol.core.CobolParser;
 import org.eclipse.lsp.cobol.core.engine.analysis.AnalysisContext;
 import org.eclipse.lsp.cobol.core.engine.analysis.EmbeddedCodeService;
 import org.eclipse.lsp.cobol.core.engine.dialects.DialectService;
 import org.eclipse.lsp.cobol.core.engine.errors.ErrorFinalizerService;
+import org.eclipse.lsp.cobol.core.engine.pipeline.Pipeline;
+import org.eclipse.lsp.cobol.core.engine.pipeline.stages.*;
+import org.eclipse.lsp.cobol.core.engine.pipeline.PipelineResult;
 import org.eclipse.lsp.cobol.core.engine.processor.AstProcessor;
-import org.eclipse.lsp.cobol.core.engine.symbols.SymbolAccumulatorService;
 import org.eclipse.lsp.cobol.core.engine.symbols.SymbolsRepository;
-import org.eclipse.lsp.cobol.core.model.tree.*;
-import org.eclipse.lsp.cobol.core.model.tree.logic.*;
-import org.eclipse.lsp.cobol.core.model.tree.logic.implicit.ImplicitVariablesProcessor;
-import org.eclipse.lsp.cobol.core.model.tree.statements.StatementNode;
-import org.eclipse.lsp.cobol.core.model.tree.variables.FileDescriptionNode;
-import org.eclipse.lsp.cobol.core.preprocessor.CopybookHierarchy;
 import org.eclipse.lsp.cobol.core.preprocessor.TextPreprocessor;
 import org.eclipse.lsp.cobol.core.preprocessor.delegates.GrammarPreprocessor;
-import org.eclipse.lsp.cobol.core.preprocessor.delegates.PreprocessorContext;
-import org.eclipse.lsp.cobol.core.semantics.CopybooksRepository;
-import org.eclipse.lsp.cobol.core.strategy.CobolErrorStrategy;
-import org.eclipse.lsp.cobol.core.visitor.CobolVisitor;
-import org.eclipse.lsp.cobol.core.visitor.ParserListener;
 import org.eclipse.lsp.cobol.service.settings.CachingConfigurationService;
 import org.eclipse.lsp.cobol.service.utils.ServerTypeUtil;
 import org.eclipse.lsp4j.Location;
@@ -86,8 +62,9 @@ import static org.eclipse.lsp.cobol.core.engine.analysis.AnalysisContext.Activit
 public class CobolLanguageEngine {
 
   private final TextPreprocessor preprocessor;
-  private final GrammarPreprocessor grammarPreprocessor;
   private final MessageService messageService;
+  private final ErrorFinalizerService errorFinalizerService;
+  private final GrammarPreprocessor grammarPreprocessor;
   private final ParseTreeListener treeListener;
   private final SubroutineService subroutineService;
   private final CachingConfigurationService cachingConfigurationService;
@@ -95,7 +72,7 @@ public class CobolLanguageEngine {
   private final AstProcessor astProcessor;
   private final SymbolsRepository symbolsRepository;
   private final EmbeddedCodeService embeddedCodeService;
-  private final ErrorFinalizerService errorFinalizerService;
+  private final Pipeline pipeline;
 
   @Inject
   public CobolLanguageEngine(
@@ -111,8 +88,9 @@ public class CobolLanguageEngine {
       EmbeddedCodeService embeddedCodeService,
       ErrorFinalizerService errorFinalizerService) {
     this.preprocessor = preprocessor;
-    this.grammarPreprocessor = grammarPreprocessor;
     this.messageService = messageService;
+    this.errorFinalizerService = errorFinalizerService;
+    this.grammarPreprocessor = grammarPreprocessor;
     this.treeListener = treeListener;
     this.subroutineService = subroutineService;
     this.cachingConfigurationService = cachingConfigurationService;
@@ -120,7 +98,13 @@ public class CobolLanguageEngine {
     this.astProcessor = astProcessor;
     this.symbolsRepository = symbolsRepository;
     this.embeddedCodeService = embeddedCodeService;
-    this.errorFinalizerService = errorFinalizerService;
+
+    this.pipeline = new Pipeline();
+    this.pipeline.add(new DialectProcessingStage(dialectService));
+    this.pipeline.add(new PreprocessorStage(grammarPreprocessor));
+    this.pipeline.add(new ParserStage(messageService, treeListener));
+    this.pipeline.add(new EmbeddedCodeStage(embeddedCodeService, treeListener));
+    this.pipeline.add(new TransformTreeStage(symbolsRepository, messageService, subroutineService, cachingConfigurationService, dialectService, astProcessor));
   }
 
   /**
@@ -145,66 +129,35 @@ public class CobolLanguageEngine {
 
     // Cleaning up
     ResultWithErrors<TextTransformations> resultWithErrors = preprocessor.cleanUpCode(documentUri, text);
-    AnalysisContext ctx = new AnalysisContext(new ExtendedSource(resultWithErrors.getResult()), analysisConfig);
+    AnalysisContext ctx = new AnalysisContext(new ExtendedSource(resultWithErrors.getResult()), analysisConfig, documentUri);
     ctx.getAccumulatedErrors().addAll(resultWithErrors.getErrors());
 
-    // Dialect processing
-    dialectService.updateDialects(analysisConfig.getDialectRegistry());
-    DialectOutcome dialectOutcome = ctx.measure(DIALECTS,
-            () -> processDialects(ctx));
+    PipelineResult<?> result = pipeline.run(ctx);
 
-    // Preprocessor (replacement, copybooks)
-    CopybooksRepository copybooksRepository = ctx.measure(PREPROCESSOR, () -> runPreprocessor(documentUri, ctx));
-    applyDialectCopybooks(copybooksRepository, dialectOutcome.getDialectNodes());
+    if (result.stopProcessing() || !(result.getData() instanceof ProcessingResult)) {
+      return new ResultWithErrors<>(
+          AnalysisResult.builder()
+              .rootNode(null)
+              .symbolTableMap(ImmutableMap.of())
+              .build(),
+          ctx.getAccumulatedErrors().stream().map(errorFinalizerService::localizeErrorMessage).collect(toList()));
+    } else {
+      ProcessingResult processingResult = (ProcessingResult) result.getData();
 
-    // Run parser
-    ParserListener listener = new ParserListener(ctx.getExtendedSource(), copybooksRepository);
-    CobolLexer lexer = new CobolLexer(CharStreams.fromString(ctx.getExtendedSource().extendedText()));
-    lexer.removeErrorListeners();
-    CommonTokenStream tokens = new CommonTokenStream(lexer);
+      ctx.measure(LATE_ERROR_PROCESSING,
+          () -> errorFinalizerService.processLateErrors(ctx, ctx.getCopybooksRepository()));
 
-    CobolParser.StartRuleContext tree = ctx.measure(PARSER,
-            () -> runParser(listener, lexer, tokens));
+      if (LOG.isDebugEnabled()) {
+        ctx.logTiming();
+      }
 
-    // Parse embedded code
-    List<Node> embeddedNodes = embeddedCodeService.generateNodes(ctx.getExtendedSource(), new ParserListener(ctx.getExtendedSource(), copybooksRepository), tree, treeListener,
-        documentUri, analysisConfig.getFeatures())
-        .unwrap(ctx.getAccumulatedErrors()::addAll);
-
-    ctx.getExtendedSource().commitTransformations();
-    // Transform parsed tree to AST
-    List<Node> syntaxTree = ctx.measure(VISITOR,
-            () -> transformAST(
-                    ctx,
-                    dialectOutcome.getDialectNodes(),
-                    copybooksRepository,
-                    tokens,
-                    tree));
-
-    addEmbeddedNodes(syntaxTree.get(0), embeddedNodes);
-
-    SymbolAccumulatorService symbolAccumulatorService = new SymbolAccumulatorService();
-    Node rootNode = ctx.measure(SYNTAX_TREE,
-            () -> {
-      Node root = processSyntaxTree(analysisConfig, symbolAccumulatorService, ctx, syntaxTree);
-      symbolsRepository.updateSymbols(symbolAccumulatorService.getProgramSymbols());
-      return root;
-    });
-
-    ctx.getAccumulatedErrors().addAll(listener.getErrors());
-    ctx.measure(LATE_ERROR_PROCESSING,
-        () -> errorFinalizerService.processLateErrors(ctx, copybooksRepository));
-
-    if (LOG.isDebugEnabled()) {
-      ctx.logTiming();
+      return new ResultWithErrors<>(
+          AnalysisResult.builder()
+              .rootNode(processingResult.getRootNode())
+              .symbolTableMap(processingResult.getSymbolTableMap())
+              .build(),
+          ctx.getAccumulatedErrors().stream().map(errorFinalizerService::localizeErrorMessage).collect(toList()));
     }
-
-    return new ResultWithErrors<>(
-        AnalysisResult.builder()
-            .rootNode(rootNode)
-            .symbolTableMap(symbolAccumulatorService.getProgramSymbols())
-            .build(),
-            ctx.getAccumulatedErrors().stream().map(errorFinalizerService::localizeErrorMessage).collect(toList()));
   }
 
   private ResultWithErrors<AnalysisResult> getErrorForIncompatibleServerTypeAndDialects(String documentUri) {
@@ -221,138 +174,4 @@ public class CobolLanguageEngine {
             ));
   }
 
-  private void addEmbeddedNodes(Node rootNode, List<Node> embeddedNodes) {
-    for (Node dialectNode : embeddedNodes) {
-      Optional<Node> nodeByPosition =
-          RangeUtils.findNodeByPosition(
-              rootNode,
-              dialectNode.getLocality().getUri(),
-              dialectNode.getLocality().getRange().getStart());
-
-      nodeByPosition.orElse(rootNode).addChild(dialectNode);
-    }
-  }
-
-  private Node processSyntaxTree(AnalysisConfig analysisConfig, SymbolAccumulatorService symbolAccumulatorService, AnalysisContext ctx, List<Node> syntaxTree) {
-    Node rootNode = syntaxTree.get(0);
-    ProcessingContext processingContext = new ProcessingContext(new ArrayList<>(), symbolAccumulatorService, ctx.getConfig().getDialectsSettings());
-    registerProcessors(analysisConfig, processingContext, symbolAccumulatorService);
-    ctx.getAccumulatedErrors().addAll(astProcessor.processSyntaxTree(processingContext, rootNode));
-    return rootNode;
-  }
-
-  private List<Node> transformAST(AnalysisContext ctx, List<Node> dialectNodes,
-                                  CopybooksRepository copybooksRepository, CommonTokenStream tokens,
-                                  CobolParser.StartRuleContext tree) {
-    CobolVisitor visitor =
-        new CobolVisitor(copybooksRepository, tokens, ctx.getExtendedSource(),
-            messageService, subroutineService, dialectNodes, cachingConfigurationService);
-    List<Node> syntaxTree = visitor.visit(tree);
-    ctx.getAccumulatedErrors().addAll(visitor.getErrors());
-    return syntaxTree;
-  }
-
-  private CobolParser.StartRuleContext runParser(ParserListener listener, CobolLexer lexer, CommonTokenStream tokens) {
-    ThreadInterruptionUtil.checkThreadInterrupted();
-    lexer.addErrorListener(listener);
-    CobolParser parser = new CobolParser(tokens);
-    parser.removeErrorListeners();
-    parser.addErrorListener(listener);
-    parser.setErrorHandler(new CobolErrorStrategy(messageService));
-    parser.addParseListener(treeListener);
-    return parser.startRule();
-  }
-
-  private CopybooksRepository runPreprocessor(String programDocumentUri, AnalysisContext ctx) {
-    List<SyntaxError> preprocessorErrors = new ArrayList<>();
-    ExtendedSource extendedSource = ctx.getExtendedSource();
-    CopybooksRepository copybooks =
-            grammarPreprocessor.preprocess(new PreprocessorContext(programDocumentUri, extendedSource, extendedSource.getMainMap(),
-                    ctx.getConfig().getCopybookConfig(), new CopybookHierarchy(), new CopybooksRepository()))
-                    .unwrap(preprocessorErrors::addAll);
-    extendedSource.commitTransformations();
-
-    ctx.getAccumulatedErrors().addAll(preprocessorErrors);
-    return copybooks;
-  }
-
-  private DialectOutcome processDialects(AnalysisContext ctx) {
-    CopybookConfig copybookConfig = ctx.getConfig().getCopybookConfig();
-    DialectProcessingContext dialectProcessingContext =
-            DialectProcessingContext.builder()
-                    .copybookConfig(copybookConfig)
-                    .programDocumentUri(ctx.getExtendedSource().getUri())
-                    .extendedSource(ctx.getExtendedSource())
-                    .build();
-    dialectProcessingContext.getExtendedSource().commitTransformations();
-
-    DialectOutcome dialectOutcome = dialectService
-                    .process(ctx.getConfig().getDialects(), dialectProcessingContext)
-                    .unwrap(ctx.getAccumulatedErrors()::addAll);
-    Set<SyntaxError> errors = new HashSet<>(ctx.getAccumulatedErrors());
-    ctx.getAccumulatedErrors().clear();
-    ctx.getAccumulatedErrors().addAll(errors);
-    return dialectOutcome;
-  }
-
-  private void registerProcessors(AnalysisConfig analysisConfig, ProcessingContext ctx, SymbolAccumulatorService symbolAccumulatorService) {
-    // Phase TRANSFORMATION
-    ProcessingPhase t = ProcessingPhase.TRANSFORMATION;
-    ctx.register(t, CompilerDirectiveNode.class, new CompilerDirectiveProcess());
-    ctx.register(t, ProgramIdNode.class, new ProgramIdProcess());
-    ctx.register(t, SectionNode.class, new SectionNodeProcessor(symbolAccumulatorService));
-    ctx.register(t, FileEntryNode.class, new FileEntryProcess());
-    ctx.register(t, FileDescriptionNode.class, new FileDescriptionProcess(symbolAccumulatorService));
-    ctx.register(t, DeclarativeProcedureSectionNode.class, new DeclarativeProcedureSectionRegister(symbolAccumulatorService));
-
-    // Phase DEFINITION
-    ProcessingPhase d = ProcessingPhase.DEFINITION;
-    ctx.register(d, ParagraphsNode.class, new DefineCodeBlock(symbolAccumulatorService));
-    ctx.register(d, SectionNameNode.class, new SectionNameRegister(symbolAccumulatorService));
-    ctx.register(d, ParagraphNameNode.class, new ParagraphNameRegister(symbolAccumulatorService));
-    ctx.register(d, ProcedureDivisionBodyNode.class, new DefineCodeBlock(symbolAccumulatorService));
-
-    // Phase POST DEFINITION
-    ctx.register(ProcessingPhase.POST_DEFINITION, SectionNode.class, new ImplicitVariablesProcessor());
-
-    // Phase USAGE
-    ProcessingPhase u = ProcessingPhase.USAGE;
-    ctx.register(u, CodeBlockUsageNode.class, new CodeBlockUsage(symbolAccumulatorService));
-    ctx.register(u, RootNode.class, new RootNodeUpdateCopyNodesByPositionInTree());
-    ctx.register(u, QualifiedReferenceNode.class, new QualifiedReferenceUpdateVariableUsage(symbolAccumulatorService));
-
-    // ENRICHMENT
-    ProcessingPhase e = ProcessingPhase.ENRICHMENT;
-    ctx.register(e, SectionNameNode.class, new SectionNameNodeEnricher(symbolAccumulatorService));
-    ctx.register(e, ParagraphNameNode.class, new ParagraphNameNodeEnricher(symbolAccumulatorService));
-    ctx.register(e, CodeBlockUsageNode.class, new CodeBlockUsageNodeEnricher(symbolAccumulatorService));
-
-    // Phase VALIDATION
-    ProcessingPhase v = ProcessingPhase.VALIDATION;
-    ctx.register(v, VariableWithLevelNode.class, new VariableWithLevelCheck());
-    ctx.register(v, StatementNode.class, new StatementValidate());
-    ctx.register(v, ElementaryNode.class, new ElementaryNodeCheck());
-    ctx.register(v, GroupItemNode.class, new GroupItemCheck());
-    ctx.register(v, ObsoleteNode.class, new ObsoleteNodeCheck());
-    ctx.register(v, StandAloneDataItemNode.class, new StandAloneDataItemCheck());
-    ctx.register(v, ProgramEndNode.class, new ProgramEndCheck());
-    ctx.register(v, CICSTranslatorNode.class, new CICSTranslatorProcessor(analysisConfig, messageService));
-    ctx.register(t, JsonParseNode.class, new JsonParseProcess(symbolAccumulatorService));
-    ctx.register(t, JsonGenerateNode.class, new JsonGenerateProcess(symbolAccumulatorService));
-
-    // Dialects
-    List<ProcessorDescription> pds = dialectService.getProcessors(analysisConfig.getDialects());
-    pds.forEach(ctx::register);
-  }
-
-  private void applyDialectCopybooks(CopybooksRepository copybooksRepository, List<Node> dialectNodes) {
-    dialectNodes.stream()
-        .filter(n -> n instanceof CopyNode)
-        .map(CopyNode.class::cast)
-        .filter(n -> n.getUri() != null)
-        .forEach(n -> {
-          copybooksRepository.addStatement(n.getName(), n.getDialect(), n.getLocality());
-          copybooksRepository.define(n.getName(), n.getDialect(), n.getNameLocation().getUri(), n.getUri());
-        });
-  }
 }
